Add kafka

+++ spring-boot-archetype/src/main/java/zbdy/controller/UserRestController.java
@@ -3,6 +3,8 @@ package zbdy.controller;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,19 +16,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import zbdy.exception.UserNotfoundException;
 import zbdy.model.User;
 
+
+//https://www.yiibai.com/spring-boot/spring_boot_building_restful_web_services.html
+//https://www.yiibai.com/spring_mvc/spring-mvc-4-restful-web-services-crud-example-resttemplate.html
+
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/users")
 public class UserRestController {
 
-	private User user = new User("zbdy", "1111");
-
-	@GetMapping()
-	public User getUser() {
-		System.out.println("------------restful----------");
-		return this.user;
-	}
+	private static Logger logger = LoggerFactory.getLogger(UserRestController.class);
 
 	private static Map<String, User> userRepo = new HashMap<>();
 	static {
@@ -37,10 +38,17 @@ public class UserRestController {
 		userRepo.put(user2.getUsername(), user2);
 	}
 
-//	@GetMapping()
-//	public ResponseEntity<Object> getUsers(){
-//		return new ResponseEntity<>(userRepo.values(), HttpStatus.OK);
-//	}
+	@GetMapping()
+	public ResponseEntity<Object> getAllUsers(){
+		return new ResponseEntity<>(userRepo.values(), HttpStatus.OK);
+	}
+	@GetMapping("/{username}")
+	public ResponseEntity<Object> getUser(@PathVariable("username") String username){
+		if(!userRepo.containsKey(username)) {
+			throw new UserNotfoundException();
+		}
+		return new ResponseEntity<>(userRepo.get(username), HttpStatus.OK);
+	}
 	
 	@PostMapping()
 	public ResponseEntity<Object> createUser(@RequestBody User user){
@@ -55,7 +63,11 @@ public class UserRestController {
 	}
 	@DeleteMapping("/{username}")
 	public ResponseEntity<Object> deleteUser(@PathVariable("username") String username){
+		if(!userRepo.containsKey(username)) {
+			throw new UserNotfoundException();
+		}
 		userRepo.remove(username);
 		return new ResponseEntity<>("User is deleted successfully", HttpStatus.OK);
 	}
 }
+
